Aggiunti ENTITY_KILLED e ROOM_CHANGED PLS: rimuovere ogni riferimento al file `EntityMoveE.hpp`
