+++ include/EntityMoveE.hpp
@@ -1,11 +1,3 @@
 #pragma once
 
-#include "RoomEvent.hpp"
-#include "Room.hpp"
-#include "geometry.h"
-class EntityMoveE : public RoomEvent {
-    public:
-        EntityMoveE(coords old_pos, coords new_pos);
-        coords data[2];
-};
-
+#include "Events.hpp"
